tests/wpt/tests/workers/interfaces/WorkerUtils/navigator/006.html

<!doctype html>
<title>navigator.onLine</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('006.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, navigator.onLine);
    this.done();
  });
});
</script>
